Initial Commit.
Creating the Creative Competition Bracket that is found here:
http://www.wrestlingassistant.com/CCB.html

I needed to implement a bracket like this for a soccer tournament
that SLCC is hosting. I like all the logic that is builtin to the
d.o/project/bracket module, and I didn't want to forfeit that and
start from scratch. So, I created an additional competition type.

Currently, you still need to customize the routing of the competition
in order to get the people going to the proper matches.

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>